refactor: Separate components into ui, macro-calculator and recipe-planner
